:recycle: move session summary serializer
