+++ api/src/certification/session-management/infrastructure/serializers/session-summary-serializer.js
