+++ api/tests/certification/session-management/unit/infrastructure/serializers/session-summary-serializer_test.js
@@ -1,5 +1,5 @@
-import * as serializer from '../../../../../lib/infrastructure/serializers/jsonapi/session-summary-serializer.js';
-import { domainBuilder, expect } from '../../../../test-helper.js';
+import * as serializer from '../../../../../../src/certification/session-management/infrastructure/serializers/session-summary-serializer.js';
+import { domainBuilder, expect } from '../../../../../test-helper.js';
 
 describe('Unit | Serializer | JSONAPI | session-summary-serializer', function () {
   describe('#serialize()', function () {
